+++ SecondPlugin/Source/Lfo.h
@@ -19,12 +19,12 @@ class LFO
         ~LFO();
 
         void reset();
-        void setSampleRate(double inSampleRate);
+        void setSampleRate(float inSampleRate);
         void process(float inRate, float inDepth, int inNumSamples);
         float* getBuffer();
 
     private:
-        double mSampleRate;
+        float mSampleRate;
         float mPhase;
         float mBuffer[maxBufferSize];
 };
